requirements.txt
apex==0.1
h5py==2.10.0
matplotlib==3.1.2
numpy==1.17.3
pandas==0.25.3
Pillow==7.1.2
pytorch-ignite==0.2.1
scipy==1.3.1
tensorboard==2.0.1
tensorboardX==1.9
tensorflow-gpu==2.0.0
torch==1.3.0
torchvision==0.4.1